make k8s version compatibility chart more scalable
Looking at the docs, we hadn't updated this chart since the v1.20 release. Adding a line with each release makes it easy to get outdated, and it gets really long if we release frequently. I updated to a format that still conveys the info, but should be easier to maintain while taking up less content area.